Conflicts:
	src/ibis/constellation/impl/Message.java
	src/ibis/constellation/impl/pool/Pool.java

+++ src/ibis/constellation/impl/pool/communication/CommunicationLayer.java
@@ -13,7 +13,7 @@ public interface CommunicationLayer {
 
     public NodeIdentifier getMaster();
 
-    public long getRank();
+    public int getRank();
 
     public int getPoolSize();
